Fix entity exports for ESM projects
Before this commit, for projects with `"type": "module"`,  anything
exported from a file in entity with `export default` was not resolvable
from the barrel file as a named export, if the importing project had
`"type": "module"`.

After this change, the interface of the index is unchanged, but ESM
projects can use it.